requirements.txt
# OPTIC: Sistema de Traducción Bidireccional de Lenguaje de Señas - Dependencias
# Ocular Powered Translation & Interpretation Core
# ===============================================================================

# Core ML y Visión Artificial
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
scipy==1.11.4
tensorflow==2.13.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0

# Procesamiento de Video y Tiempo Real
ffmpeg-python==0.2.0
moviepy==1.0.3
av==10.0.0

# Interfaz Gráfica
customtkinter==5.2.2
pillow==10.2.0
tkinter-tooltip==2.0.0

# Audio y Multimedia
pygame==2.5.2
sounddevice==0.4.6
librosa==0.10.1

# Comunicación Web y APIs
requests==2.31.0
urllib3==2.0.7
fastapi==0.104.1
uvicorn==0.24.0

# Manejo de Datos
pandas==2.0.3
python-dateutil==2.8.2
h5py==3.10.0

# Utilidades
difflib2==0.1.1
typing-extensions==4.8.0
tqdm==4.65.0

# Texto a Voz (TTS)
pyttsx3==2.90
gTTS==2.4.0

# Desarrollo y Debugging
matplotlib==3.7.1
jupyter==1.0.0
pytest==7.4.3

# Opcional: Para desarrollo y debugging
# matplotlib==3.7.2
# jupyter==1.0.0

# Aprendizaje profundo y machine learning
scikit-learn==1.3.0

# Procesamiento de datos
seaborn==0.12.2

# Base de datos
sqlite3  # Incluido en Python estándar

# Utilidades
logging  # Incluido en Python estándar
argparse  # Incluido en Python estándar
json  # Incluido en Python estándar
os  # Incluido en Python estándar
time  # Incluido en Python estándar
threading  # Incluido en Python estándar
typing  # Incluido en Python estándar
collections  # Incluido en Python estándar
dataclasses  # Incluido en Python estándar 